+++ CHANGELOG.md
@@ -13,6 +13,14 @@ Types of changes
 
 ## [Unreleased]
 
+### Added
+
+- allocator-free version of the TZif timezone (no heap memory required)
+
+### Fixed
+
+- tests in tzif.zig; were testing std lib tzif reader, now test this implementation
+
 ## 2025-03-20, v0.6.1
 
 ### Added